acs-language/extract-information/10_requirements.txt
azure-ai-textanalytics==5.2.0b1